add authentication to resolvers; login now available

+++ server/schemas/resolvers.js
@@ -1,7 +1,23 @@
-const { User, Rest } = require('../models');
+const { AuthenticationError } = require('apollo-server-express');
+const { User, Rest, Dish } = require('../models');
+const { signToken } = require('../utils/auth');
 
 const resolvers = {
   Query: {
+    // get logged in user
+    me: async (parent, args, context) => {
+      if (context.user) {
+        const userData = await User.findOne({ _id: context.user._id }).select(
+          '-__v -password'
+        );
+
+        return userData;
+      }
+
+      throw new AuthenticationError('Not logged in');
+    },
+
+    // get all users
     users: async () => {
       return await User.find()
         .select('-__v -password')
@@ -9,6 +25,7 @@ const resolvers = {
         .populate('comments');
     },
 
+    // get user by username
     user: async (parent, { username }) => {
       return await User.findOne({ username })
         .select('-__v -password')
@@ -16,45 +33,111 @@ const resolvers = {
         .populate('comments');
     },
 
+    // get all restaurants
     restaurants: async () => {
       return await Rest.find().select('-__v').populate('comments');
     },
 
+    // get restaurant by name
     restaurant: async (parent, { restName }) => {
       return await Rest.find({ restName }).select('-__v').populate('comments');
     }
   },
 
   Mutation: {
+    // add new user
     addUser: async (parent, args) => {
       const user = await User.create(args);
+      const token = signToken(user);
 
       return user;
     },
 
-    followUser: async (parent, { userId, userToFollowId }) => {
-      const user = await User.findById(userId);
-      const userToFollow = await User.findById(userToFollowId);
+    // login user
+    login: async (parent, { email, password }) => {
+      const user = await User.findOne({ email });
 
-      user.following.push(userToFollow);
-      user.save();
+      if (!user) {
+        throw new AuthenticationError('Incorrect credentials');
+      }
 
-      userToFollow.followers.push(user);
-      userToFollow.save();
+      const correctPw = await user.isCorrectPassword(password);
 
-      return user;
+      if (!correctPw) {
+        throw new AuthenticationError('Incorrect credentials');
+      }
+
+      const token = signToken(user);
+      return { token, user };
     },
 
-    addRest: async (parent, args) => {
-      const rest = await Rest.create(args);
+    // follow another user
+    followUser: async (parent, { userToFollowId }, context) => {
+      if (context.user) {
+        const user = await User.findById(context.user._id);
+        const userToFollow = await User.findById(userToFollowId);
 
-      return rest;
+        user.following.push(userToFollow);
+        user.save();
+
+        userToFollow.followers.push(user);
+        userToFollow.save();
+
+        return user;
+      }
+
+      throw new AuthenticationError('Not logged in');
     },
 
-    deleteRest: async (parent, { restId }) => {
-      const rest = await Rest.findByIdAndDelete(restId);
+    // unfollow another user
+    unfollowUser: async (parent, { userToUnfollowId }, context) => {
+      if (context.user) {
+        const user = await User.findById(context.user._id);
+        const userToUnfollow = await User.findById(userToUnfollowId);
 
-      return rest;
+        user.following.pull(userToUnfollow);
+        user.save();
+
+        userToUnfollow.followers.pull(user);
+        userToUnfollow.save();
+
+        return user;
+      }
+
+      throw new AuthenticationError('Not logged in');
+    },
+
+    // add a new restaurant
+    addRest: async (parent, args, context) => {
+      if (context.user) {
+        const rest = await Rest.create(args);
+
+        return rest;
+      }
+
+      throw new AuthenticationError('You need to be logged in!');
+    },
+
+    // delete a restaurant
+    deleteRest: async (parent, { restId }, context) => {
+      if (context.user) {
+        const rest = await Rest.findByIdAndDelete(restId);
+
+        return rest;
+      }
+
+      throw new AuthenticationError('You need to be logged in');
+    },
+
+    // add a new dish
+    addDish: async (parent, args, context) => {
+      if (context.user) {
+        const dish = await Dish.create(args);
+
+        return dish;
+      }
+
+      throw new AuthenticationError('You need to be logged in!');
     }
   }
 };
